+++ cli/src/main/resources/logback.xml
