+++ cli/src/main/scala/commandcenter/cli/CliCommand.scala
