+++ cli/src/main/scala/commandcenter/cli/Main.scala
