+++ cli/src/main/scala/commandcenter/cli/message/CCRequest.scala
