+++ cli/src/main/scala/commandcenter/cli/message/CCResponse.scala
