+++ cli/src/main/scala/commandcenter/cli/zmq/ZPublisher.scala
