+++ cli/src/main/scala/commandcenter/cli/zmq/ZSubscriber.scala
