+++ emulator-swing/plugins/PLUGINS.md
