+++ emulator-swing/src/main/scala/commandcenter/emulator/swing/event/KeyboardShortcutUtil.scala
@@ -1,4 +1,4 @@
-package commandcenter.daemon.event
+package commandcenter.emulator.swing.event
 
 import java.awt.event.{ InputEvent, KeyEvent }
