+++ emulator-swing/src/main/scala/commandcenter/emulator/swing/ui/ZTextField.scala
@@ -1,4 +1,4 @@
-package commandcenter.daemon.ui
+package commandcenter.emulator.swing.ui
 
 import java.awt.event.{ KeyEvent, KeyListener }
